+++ tests/test_channel_props/test_choi_rank.py
@@ -27,8 +27,7 @@ def test_choi_rank_list_kraus():
 
 def test_choi_rank_choi_matrix():
     """Verify Choi matrix of the swap operator map gives correct Choi rank."""
-    choi_matrix = np.array([[1, 0, 0, 0], [0, 0, 1, 0],
-                            [0, 1, 0, 0], [0, 0, 0, 1]])
+    choi_matrix = np.array([[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]])
     np.testing.assert_equal(choi_rank(choi_matrix), 4)
 
 
@@ -36,7 +35,7 @@ def test_choi_bad_input():
     """Verify that a bad input (such as a string which still passes
     with `numpy.linalg.matrix_rank`) raises an error"""
     with pytest.raises(ValueError, match="Not a valid"):
-        bad_input = 'string'
+        bad_input = "string"
         choi_rank(bad_input)